Add support for LLVM source-based coverage on Windows
- Updated .gitignore to include build-coverage directory.
- Enhanced CMakeLists.txt to introduce YSE_LLVM_COVERAGE option for LLVM coverage instrumentation on Windows.
- Added coverage-windows preset in CMakePresets.json for building with LLVM coverage.
- Modified various source files to suppress unused parameter warnings and improve code clarity.
- Updated yse.py to handle coverage reporting for both Linux and Windows, utilizing llvm-profdata and llvm-cov on Windows.
- Created coverage-llvm.json file for storing coverage report output.

+++ YseEngine/dsp/interpolate4.cpp
@@ -24,7 +24,7 @@ YSE::DSP::buffer * YSE::DSP::interpolate4::source() {
 
 YSE::DSP::interpolate4 & YSE::DSP::interpolate4::onset(Int value) {
   if (data) {
-    Clamp(value, 0, data->getLength() - 1);
+    Clamp(value, 0, static_cast<int>(data->getLength()) - 1);
     parmOnset.store(value);
   }
   return *this;
